Trim Apple Pencil PR: drop barrel button and extra files
Remove barrel-button wire path and UIPencilInteraction. Delete
PencilPhase.swift and InputCapture.swift; use string phase literals
and inline InputCaptureEngine in OpenSidecarPhoneApp.swift. Restore
upstream normalized() body and InputInjector line comments.
