--- Shared/PencilPhase.swift
@@ -1,6 +1,0 @@
-import Foundation
-
-/// Phases on the `"pencil"` control message (`phase` field).
-enum PencilPhase: String {
-    case down, move, up, hover
-}
